% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFingerprints.R
\name{compareFingerprints}
\alias{compareFingerprints}
\title{Compare Structure Fingerprints to NCI DTP Compounds}
\usage{
compareFingerprints(ids = NULL, smiles = NULL, fpType = "standard",
  verbose = TRUE, fingerprint.list = NULL)
}
\arguments{
\item{ids}{a vector of IDs corresponding to structures}

\item{smiles}{a vector of strings SMILES structures}

\item{fpType}{the type of fingerprint to be used; uses the RCDK 
get.fingerprint() (default: standard)}

\item{verbose}{a boolean whether to display debugging information}

\item{fingerprint.list}{a list of fingerprints generated with \code{getFingerprintList}}
}
\value{
a sorted named vector of Tanimoto distances
}
\description{
This function compares the first SMILES structure to all other structures given.
}
\examples{
drugAnnot <- as(featureData(getAct(rcellminerData::drugData)), "data.frame")
ids <- head(drugAnnot$NSC)
smiles <- head(drugAnnot$SMILES)
fingerprintList <- getFingerprintList(ids, smiles)
compareFingerprints(fingerprint.list=fingerprintList)

}
\seealso{
rcdk::get.fingerprint
}
\concept{
rcellminer
}
