% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crossCors.R
\name{crossCorsSpearman}
\alias{crossCorsSpearman}
\title{Calculate Spearman's correlations with between rows of input matrices}
\usage{
crossCorsSpearman(X, Y = NULL)
}
\arguments{
\item{X}{a matrix or data.frame}

\item{Y}{a matrix or data.frame}
}
\value{
a list containing matrices of pairwise Spearman's correlations and
their p-values between rows of the input matrices or dataframes.
}
\description{
Calculate Spearman's correlations with between rows of input matrices
}
\examples{
\dontrun{
crossCorsSpearman(drugActData[c("94600"), ], drugActData[c("727625", "670655"), ])
}
}
\concept{
rcellminer
}

