% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Drug_MOA_Key.R
\docType{data}
\name{Drug_MOA_Key}
\alias{Drug_MOA_Key}
\title{A data frame with descriptive information for all compound mechanism of action (MOA)
abbreviations used in CellMiner.}
\description{
A data frame with descriptive information for all compound mechanism of action (MOA)
abbreviations used in CellMiner.
}
\concept{
rcellminer
}
\keyword{data}

