% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugData-class.R
\docType{class}
\name{DrugData-class}
\alias{.DrugData}
\alias{DrugData-class}
\title{An S4 class to represent drug activity and related data recorded for a set of
biological samples.}
\arguments{
\item{...}{Other possible parameters.}
}
\description{
An S4 class to represent drug activity and related data recorded for a set of
biological samples.
}
\section{Slots}{

\describe{
\item{\code{act}}{An eSet object containing drug activity data across a set of biological samples.}

\item{\code{repeatAct}}{An eSet object containing repeat drug activity experiment data with
respect to the same samples associated with act.}

\item{\code{sampleData}}{A MIAxE object capturing sample and other data set information.}
}}
\concept{
rcellminerData
}

