% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\docType{methods}
\name{DrugData,eSet,eSet,MIAxE-method}
\alias{DrugData,eSet,eSet,MIAxE-method}
\title{Returns a DrugData object.}
\usage{
\S4method{DrugData}{eSet,eSet,MIAxE}(act, repeatAct, sampleData, ...)
}
\arguments{
\item{act}{An eSet object containing drug activity data across a set of biological samples.}

\item{repeatAct}{An eSet object containing repeat drug activity experiment data with
respect to the same samples associated with act.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}

\item{...}{Other possible parameters.}
}
\value{
A DrugData object.
}
\description{
Returns a DrugData object.
}
\concept{
rcellminerData
}

