% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{DrugData}
\alias{DrugData}
\title{Returns a DrugData object.}
\usage{
DrugData(act, repeatAct, sampleData, ...)
}
\arguments{
\item{act}{An eSet object containing drug activity data across a set of biological samples.}

\item{repeatAct}{An eSet object containing repeat drug activity experiment data with
respect to the same samples associated with act.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}

\item{...}{Other possible parameters.}
}
\value{
A DrugData object.
}
\description{
Returns a DrugData object.
}
\concept{rcellminerData}
