% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{fingerprintList}
\alias{fingerprintList}
\title{Molecular Fingerprint List}
\description{
Molecular Fingerprint List
}
\references{
\url{http://discover.nci.nih.gov/cellminer}
}
\author{
Augustin Luna \email{augustin AT mail.nih.gov}
}
\concept{rcellminer}
\keyword{data}
