% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\docType{methods}
\name{getAct,DrugData-method}
\alias{getAct,DrugData-method}
\title{Returns an eSet object with drug activity data.}
\usage{
\S4method{getAct}{DrugData}(object)
}
\arguments{
\item{object}{DrugData object for which drug activity data is to be returned.}
}
\value{
An eSet object with drug activity data.
}
\description{
Returns an eSet object with drug activity data.
}
\concept{
rcellminerData
}

