% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getAct}
\alias{getAct}
\title{Returns an eSet object with drug activity data.}
\usage{
getAct(object, ...)
}
\arguments{
\item{object}{Object for which drug activity data is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
An eSet object with drug activity data.
}
\description{
Returns an eSet object with drug activity data.
}
\concept{rcellminerData}
