% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{getAllFeatureData,MolData-method}
\alias{getAllFeatureData,MolData-method}
\title{Returns a list of feature data matrices.}
\usage{
\S4method{getAllFeatureData}{MolData}(object)
}
\arguments{
\item{object}{MolData object for which a list of feature data matrices is to be returned.}
}
\value{
A list of feature data matrices.
}
\description{
Returns a list of feature data matrices.
}
\concept{
rcellminerData
}

