% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getAllFeatureData}
\alias{getAllFeatureData}
\title{Returns a list of feature data matrices.}
\usage{
getAllFeatureData(object, ...)
}
\arguments{
\item{object}{Object for which a list of feature data matrices is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
A list of feature data matrices.
}
\description{
Returns a list of feature data matrices.
}
\concept{rcellminerData}
