% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnQuantiles.R
\name{getColumnQuantiles}
\alias{getColumnQuantiles}
\title{Calculate quantile for the columns in a matrix}
\usage{
getColumnQuantiles(X, prob, naRm = FALSE, onlyNonzeroVals = FALSE)
}
\arguments{
\item{X}{the matrix}

\item{prob}{a numeric probablity}

\item{naRm}{a boolean, whether to remove NAs}

\item{onlyNonzeroVals}{a boolean, whether to only include non-zero values}
}
\value{
a vector of quantiles
}
\description{
Calculate quantile for the columns in a matrix
}
\examples{
getColumnQuantiles(matrix(1:25, nrow=5), prob = 0.5)

}
\concept{
rcellminer
}

