% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugMoaList.R
\name{getDrugMoaList}
\alias{getDrugMoaList}
\title{Get a list of applicable MOA strings for a drug.}
\usage{
getDrugMoaList(nsc, moaToCompoundListMap = NULL)
}
\arguments{
\item{nsc}{An NSC string.}

\item{moaToCompoundListMap}{A named list of character vectors, with each name
indicating an MOA class, and its corresponding character vector specifying MOA-associated
drugs. If unspecified, this is constructed based on MOA information provided by CellMiner.}
}
\value{
A character vector giving all MOA classes for the drug.
}
\description{
Get a list of applicable MOA strings for a drug.
}
\details{
LINK TO MOAs?
}
\examples{
getDrugMoaList("754365")

}
\concept{
rcellminer
}
