% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{getESetList,MolData-method}
\alias{getESetList,MolData-method}
\title{Returns a list of eSet objects.}
\usage{
\S4method{getESetList}{MolData}(object)
}
\arguments{
\item{object}{MolData object for which a list of eSet objects is to be returned.}
}
\value{
A list of eSet objects.
}
\description{
Returns a list of eSet objects.
}
\concept{
rcellminerData
}

