% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getESetList}
\alias{getESetList}
\title{Returns a list of eSet objects.}
\usage{
getESetList(object, ...)
}
\arguments{
\item{object}{Object for which a list of eSets is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
A list of eSet objects.
}
\description{
Returns a list of eSet objects.
}
\concept{
rcellminerData
}

