% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\name{getFeatureAnnot,DrugData-method}
\alias{getFeatureAnnot,DrugData-method}
\title{Returns a list of data frames with feature information.}
\usage{
\S4method{getFeatureAnnot}{DrugData}(object)
}
\arguments{
\item{object}{DrugData object for which feature data is to be returned.}
}
\value{
A named list of data frames with feature information for drugs
 and drug repeat experiments.
}
\description{
Returns a list of data frames with feature information.
}
\concept{rcellminerData}
