% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{getFeatureAnnot,MolData-method}
\alias{getFeatureAnnot,MolData-method}
\title{Returns a list of data frames with feature information.}
\usage{
\S4method{getFeatureAnnot}{MolData}(object)
}
\arguments{
\item{object}{MolData object for which feature data is to be returned.}
}
\value{
A named list of data frames with feature information for 
 available molecular data types.
}
\description{
Returns a list of data frames with feature information.
}
\concept{
rcellminerData
}
