% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getFeatureAnnot}
\alias{getFeatureAnnot}
\title{Returns a list of data frames with feature information.}
\usage{
getFeatureAnnot(object, ...)
}
\arguments{
\item{object}{Object for which feature data is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
A list of data frames with feature information.
}
\description{
Returns a list of data frames with feature information.
}
\concept{rcellminerData}
