% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFingerprintList.R
\name{getFingerprintList}
\alias{getFingerprintList}
\title{Get a list of fingerprints for a set of compounds}
\usage{
getFingerprintList(ids, smiles, fpType = "standard", verbose = TRUE)
}
\arguments{
\item{ids}{a vector of IDs corresponding to structures}

\item{smiles}{a vector of strings SMILES structures}

\item{fpType}{the type of fingerprint to be used; uses the RCDK
get.fingerprint() (default: standard)}

\item{verbose}{show debugging output}
}
\value{
a list of fingerprints returned from RCDK where the list names are
  the IDs
}
\description{
Get a list of fingerprints for a set of compounds
}
\examples{
drugAnnot <- as(featureData(getAct(rcellminerData::drugData)), "data.frame")
ids <- head(drugAnnot$NSC)
smiles <- head(drugAnnot$SMILES)
fingerprintList <- getFingerprintList(ids, smiles)

\dontrun{
# All fingerprints
ids <- drugAnnot$NSC
smiles <- drugAnnot$SMILES
fingerprintList <- getFingerprintList(ids, smiles)
save(fingerprintList, file="fingerprintList.RData")
}

}
\seealso{
rcdk::get.fingerprint
}
\concept{
rcellminer
}
