% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinDrugActivityRepeatCor.R
\name{getMinDrugActivityRepeatCor}
\alias{getMinDrugActivityRepeatCor}
\title{Returns a table indicating, for each compound in a specified set, the least significant 
correlation and associated p-value between its replicate experiments.}
\usage{
getMinDrugActivityRepeatCor(nscSet)
}
\arguments{
\item{nscSet}{a character vector specifying NSC identifier(s) for compound(s) of interest.}
}
\value{
a dataframe containing the following columns:
NSC, cor, pval
}
\description{
Returns a table indicating, for each compound in a specified set, the least significant 
correlation and associated p-value between its replicate experiments.
}
\examples{
nscSet <- c("123528", "339316")
repExpCorTab <- getMinDrugActivityRepeatCor(nscSet)

}
\concept{
rcellminer
}

