% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMoaStr.R
\name{getMoaStr}
\alias{getMoaStr}
\title{Get MOA string}
\usage{
getMoaStr(nscStr)
}
\arguments{
\item{nscStr}{an NSC string}
}
\value{
a comma-delimited string with MOA
}
\description{
Get MOA string
}
\details{
LINK TO MOAs?
}
\examples{
getMoaStr("94600")
}
\concept{
rcellminer
}

