% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMoaToCompounds.R
\name{getMoaToCompounds}
\alias{getMoaToCompounds}
\title{Get a named list mapping MOA classes to associated compound sets.}
\usage{
getMoaToCompounds()
}
\value{
a named list mapping MOA classes to associated compound sets
(each represented as a character vector).
}
\description{
Get a named list mapping MOA classes to associated compound sets.
}
\examples{
moaToCompounds <- getMoaToCompounds()
}
\concept{
rcellminer
}

