% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMolDataMatrices.R
\name{getMolDataMatrices}
\alias{getMolDataMatrices}
\title{Returns a list of molecular data type matrices, with rownames in 
each matrix prefixed with a data type abbreviation.}
\usage{
getMolDataMatrices(molDataMats = NULL)
}
\arguments{
\item{molDataMats}{A named list of molecular data type matrices with feature data 
specified along the rows, and feature names indicated in the row names.}
}
\value{
a list containing molecular data type matrices, with rownames 
in each matrix prefixed with a data type abbreviation, e.g., 'exp' for
mRNA expression, etc.  The matrix-specific data type abbreviations are
derived from the names of molDataMats.
}
\description{
Returns a list of molecular data type matrices, with rownames in 
each matrix prefixed with a data type abbreviation.
}
\examples{
molDataMats <- getMolDataMatrices()

}
\concept{
rcellminer
}
