% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMolDataType.R
\name{getMolDataType}
\alias{getMolDataType}
\title{Get the molecular data type prefixes for a set of features.}
\usage{
getMolDataType(features, prefixLen = 3)
}
\arguments{
\item{features}{A vector of features.}

\item{prefixLen}{The length of the molecular data type prefix.}
}
\value{
A character vector of molecular data type prefixes.

#' @examples 
getMolDataType(c("expTP53", "copMDM2", "mutCHEK2", "mutBRAF"))
}
\description{
Get the molecular data type prefixes for a set of features.
}
\concept{
rcellminer
}
