% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumMissingLines.R
\name{getNumMissingLines}
\alias{getNumMissingLines}
\title{Returns a vector indicating the number of NCI-60 cell lines with missing activity data
for set of compounds.}
\usage{
getNumMissingLines(nscSet)
}
\arguments{
\item{nscSet}{a character vector specifying NSC identifier(s) for compound(s) of interest.}
}
\value{
a numeric vector indicating the number of NCI-60 cell lines with missing
activity data, indexed by the identifiers in nscSet.
}
\description{
Returns a vector indicating the number of NCI-60 cell lines with missing activity data
for set of compounds.
}
\examples{
nscSet <- c("1", "17", "89", "609699")
getNumMissingLines(nscSet)
}
\concept{
rcellminer
}

