% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\docType{methods}
\name{getRepeatAct,DrugData-method}
\alias{getRepeatAct,DrugData-method}
\title{Returns an eSet object with drug repeat activity experiment data.}
\usage{
\S4method{getRepeatAct}{DrugData}(object)
}
\arguments{
\item{object}{DrugData object for which drug repeat activity experiment data
is to be returned.}
}
\value{
An eSet object with drug repeat activity experiment data.
}
\description{
Returns an eSet object with drug repeat activity experiment data.
}
\concept{
rcellminerData
}

