% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getRepeatAct}
\alias{getRepeatAct}
\title{Returns an eSet object with drug repeat activity experiment data.}
\usage{
getRepeatAct(object, ...)
}
\arguments{
\item{object}{Object for which drug repeat activity experiment data is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
An eSet object with drug repeat activity experiment data.
}
\description{
Returns an eSet object with drug repeat activity experiment data.
}
\concept{rcellminerData}
