% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\docType{methods}
\name{getSampleData,DrugData-method}
\alias{getSampleData,DrugData-method}
\title{Returns a data frame with sample information.}
\usage{
\S4method{getSampleData}{DrugData}(object)
}
\arguments{
\item{object}{DrugData object for which sample data is to be returned.}
}
\value{
A data frame with sample information.
}
\description{
Returns a data frame with sample information.
}
\concept{
rcellminerData
}

