% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{getSampleData,MolData-method}
\alias{getSampleData,MolData-method}
\title{Returns a data frame with sample information.}
\usage{
\S4method{getSampleData}{MolData}(object)
}
\arguments{
\item{object}{MolData object for which sample data is to be returned.}
}
\value{
A data frame with sample information.
}
\description{
Returns a data frame with sample information.
}
\concept{
rcellminerData
}

