% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getSampleData}
\alias{getSampleData}
\title{Returns a data frame with sample information.}
\usage{
getSampleData(object, ...)
}
\arguments{
\item{object}{Object for which sample data is to be returned.}

\item{...}{Other possible parameters.}
}
\value{
A data frame with sample information.
}
\description{
Returns a data frame with sample information.
}
\concept{
rcellminerData
}
