% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmiles.R
\name{getSmiles}
\alias{getSmiles}
\title{Get the SMILES strings for a set of NSC identifiers.}
\usage{
getSmiles(nscSet)
}
\arguments{
\item{nscSet}{A character vector of NSC strings}
}
\value{
A named character vector indicating the SMILES string for
each NSC in nscSet (or NA if no structural information is available).
}
\description{
Get the SMILES strings for a set of NSC identifiers.
}
\examples{
nscSet <- c("609699", "94600")
getSmiles(nscSet)

}
\concept{rcellminer}
