% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasMoa.R
\name{hasMoa}
\alias{hasMoa}
\title{Check if NSC has Mechanism of Action (MOA) Annotation}
\usage{
hasMoa(nsc)
}
\arguments{
\item{nsc}{a string, an NSC identifier}
}
\value{
a boolean whether the NSC has an MOA
}
\description{
Check if NSC has Mechanism of Action (MOA) Annotation
}
\examples{
hasMoa("754365")

}
\concept{rcellminer}
