% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugData-methods.R
\docType{methods}
\name{initialize,DrugData-method}
\alias{initialize,DrugData-method}
\title{Returns a DrugData object.}
\usage{
\S4method{initialize}{DrugData}(.Object, act, repeatAct, sampleData)
}
\arguments{
\item{.Object}{An object: see "new()" documentation in "methods" package.}

\item{act}{An eSet object containing drug activity data across a set of biological samples.}

\item{repeatAct}{An eSet object containing repeat drug activity experiment data with
respect to the same samples associated with act.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}
}
\value{
A DrugData object.
}
\description{
Returns a DrugData object.
}
\note{
Seems to be required for definition of a constructor.
}
\concept{
rcellminerData
}
