% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{initialize,MolData-method}
\alias{initialize,MolData-method}
\title{Returns a MolData object.}
\usage{
\S4method{initialize}{MolData}(.Object, eSetList, sampleData)
}
\arguments{
\item{.Object}{An object: see "new()" documentation in "methods" package.}

\item{eSetList}{A list of eSet objects for a common set of samples.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}
}
\value{
A MolData object.
}
\description{
Returns a MolData object.
}
\concept{
rcellminerData
}

