% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isPublic.R
\name{isPublic}
\alias{isPublic}
\title{Check if an NSC ID is public}
\usage{
isPublic(nscs)
}
\arguments{
\item{nscs}{a vector of NSC string IDs}
}
\value{
a vector of boolean values of whether each NSC is public
}
\description{
Check if an NSC ID is public
}
\examples{
isPublic("-1")
isPublic(c("-1", "609699"))
}
\concept{
rcellminer
}

