% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellminerPlotInfo.R
\name{loadCellminerPlotInfo}
\alias{loadCellminerPlotInfo}
\title{Returns data to plot CellMiner plots}
\usage{
loadCellminerPlotInfo(returnDf = FALSE)
}
\arguments{
\item{returnDf}{a boolean if a data.frame with all information (default: FALSE)}
}
\value{
a vector of colors as strings or a data.frame with dataType, label, xMin, xMax
}
\description{
Returns data to plot CellMiner plots
}
\examples{
loadCellminerPlotInfo()  
  
}
\concept{rcellminer}
