% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadNciColorSet.R
\name{loadNciColorSet}
\alias{loadNciColorSet}
\title{Returns a 60-element color set that matches the color set used on
http://discover.nci.nih.gov/}
\usage{
loadNciColorSet(returnDf = FALSE)
}
\arguments{
\item{returnDf}{a boolean if a data.frame with tissue names and abbreviations
should be returned (default: FALSE)}
}
\value{
a vector of colors as strings or a data.frame with tissues, tissue
  abbreviations, cell line abbreviations and colors
}
\description{
Returns a 60-element color set that matches the color set used on
http://discover.nci.nih.gov/
}
\examples{
loadNciColorSet()
}
\concept{
rcellminer
}

