% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolData-class.R
\docType{class}
\name{MolData-class}
\alias{MolData-class}
\alias{.MolData}
\title{An S4 class to represent molecular data recorded for a set of biological samples.}
\arguments{
\item{...}{Other possible parameters.}
}
\description{
An S4 class to represent molecular data recorded for a set of biological samples.
}
\section{Slots}{

\describe{
\item{\code{eSetList}}{A list of eSet objects for a common set of samples.}

\item{\code{sampleData}}{A MIAxE object capturing sample and other data set information.}
}}

\concept{rcellminerData}
