% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{MolData,list,MIAxE-method}
\alias{MolData,list,MIAxE-method}
\title{Returns a MolData object.}
\usage{
\S4method{MolData}{list,MIAxE}(eSetList, sampleData, ...)
}
\arguments{
\item{eSetList}{A list of eSet objects for a common set of samples.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}

\item{...}{Other possible parameters.}
}
\value{
A MolData object.
}
\description{
Returns a MolData object.
}
\concept{
rcellminerData
}

