% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{MolData}
\alias{MolData}
\title{Returns a MolData object.}
\usage{
MolData(eSetList, sampleData, ...)
}
\arguments{
\item{eSetList}{A list of eSet objects for a common set of samples.}

\item{sampleData}{A MIAxE object capturing sample and other data set information.}

\item{...}{Other possible parameters.}
}
\value{
A MolData object.
}
\description{
Returns a MolData object.
}
\concept{
rcellminerData
}
