% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parCorPatternComparison.R
\name{parCorPatternComparison}
\alias{parCorPatternComparison}
\title{Compare an input pattern against a set of patterns, excluding the effect
of a fixed pattern.}
\usage{
parCorPatternComparison(x, Y, z)
}
\arguments{
\item{x}{An N element input pattern specified as either a vector or a
1 x N matrix or data frame.}

\item{Y}{A k x N matrix with k patterns for comparison with the input pattern x
specified along the rows, with rownames set appropriately.}

\item{z}{An N element  pattern specified as either a vector or a
1 x N matrix or data frame. This is the pattern whose effect (with respect to
a linear model) is to be excluded when comparing x with each row entry of Y.}
}
\value{
A data frame with pattern comparison results (ordered by PARCOR):
NAME: Name of entry in Y being compared.
PARCOR: Partial correlation between x and the entry in Y with respect to z.
PVAL: p-value.
}
\description{
Compare an input pattern against a set of patterns, excluding the effect
of a fixed pattern.
}
\examples{
x <- exprs(getAct(rcellminerData::drugData))["609699", ]
Y <- rcellminer::getAllFeatureData(rcellminerData::molData)[["exp"]][1:100, ]
z <- rcellminer::getAllFeatureData(rcellminerData::molData)[["exp"]]["SLFN11", ]
results <- parCorPatternComparison(x, Y, z)

}
\concept{
rcellminer
}

