% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/passRuleOf5.R
\name{passRuleOf5}
\alias{passRuleOf5}
\title{Checks if SMILES passes Lipinski's Rule of 5}
\usage{
passRuleOf5(smiles, acceptableViolations = 0, verbose = FALSE)
}
\arguments{
\item{smiles}{a string, the SMILES structure to be checked}

\item{acceptableViolations,}{a number, the number of acceptable rule violations (default: 0)}

\item{verbose}{a boolean, whether to write out the failing criteria (default: FALSE)}
}
\value{
a boolean, whether the SMILES passes the criteria
}
\description{
Checks if SMILES passes Lipinski's Rule of 5
}
\details{
Uses RCDK: org.openscience.cdk.qsar.descriptors.molecular.RuleOfFiveDescriptor
}
\examples{
# Docetaxel
passRuleOf5("CC1=C2C(C(=O)C3(C(CC4C(C3C(C(C2(C)C)(CC1OC(=O)C(C(C5=CC=CC=C5)NC
(=O)OC(C)(C)C)O)O)OC(=O)C6=CC=CC=C6)(CO4)OC(=O)C)O)C)O", verbose=TRUE)

# Gemcitabine
passRuleOf5("C1=CN(C(=O)N=C1N)C2C(C(C(O2)CO)O)(F)F", verbose=TRUE)
}
\concept{
rcellminer
}

