% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passRuleOf5FromNsc.R
\name{passRuleOf5FromNsc}
\alias{passRuleOf5FromNsc}
\title{Checks if NSC passes Lipinski's Rule of 5}
\usage{
passRuleOf5FromNsc(nsc, acceptableViolations = 0, verbose = FALSE)
}
\arguments{
\item{nsc}{a string, the NSC identifier to be checked}

\item{acceptableViolations, }{a number, the number of acceptable rule violations (default: 0)}

\item{verbose}{a boolean, whether to write out the failing criteria (default: FALSE)}
}
\value{
a boolean or NA, whether the NSC passes the criteria or NA if the NSC had no structure
}
\description{
Checks if NSC passes Lipinski's Rule of 5
}
\details{
Uses RCDK: org.openscience.cdk.qsar.descriptors.molecular.RuleOfFiveDescriptor
}
\examples{
passRuleOf5FromNsc("94600", verbose=TRUE)

}
\concept{
rcellminer
}
