% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCellMiner.R
\name{plotCellMiner}
\alias{plotCellMiner}
\title{Description: Produces CellMiner-like plots in R}
\usage{
plotCellMiner(drugAct, molData, plots, nsc = NULL, gene = NULL,
  sub = NULL, xLimits = c(-3, 3), xLabel = "Z-scores", verbose = FALSE)
}
\arguments{
\item{drugAct}{a matrix of drug activity values (cell lines as columns, drug entries as rows)}

\item{molData}{a list of matricies a molecular}

\item{plots}{a vector of characters denoting the plots to include and the order (e.g. c("mut", "drug", "cop").
Currently, supported entries mutations (mut), drug activities (drug), copy number variations (cop)}

\item{nsc}{a string NSC ID that will be plotted when a "drug" entry appears in the plots vector}

\item{gene}{a string HUGO gene symbol for which the "mut", "cop", or "exp" plots will be produced if in plots vector}

\item{sub}{a vector of strings with sub-titles for each plot}

\item{xLimits}{a 2 number vector with the the minimum and maximum X-axis values (default: c(-3, 3))}

\item{xLabel}{a string for the default X-axis label}

\item{verbose}{a boolean to show debugging information}
}
\value{
None
}
\description{
Description: Produces CellMiner-like plots in R
}
\examples{
drugAct <- exprs(getAct(rcellminerData::drugData))
molDataMats <- getMolDataMatrices()
plotDataMats <- molDataMats[c("exp", "cop", "mut")]
plotCellMiner(drugAct, plotDataMats,
plots=c("mut", "drug", "cop"), nsc="94600", gene="TP53")
}
\author{
Augustin Luna <augustin AT mail.nih.gov>
}
\concept{
rcellminer
}

