% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellMiner2D.R
\name{plotCellMiner2D}
\alias{plotCellMiner2D}
\title{Make a simple 2d plot using two variables with ggplot2}
\usage{
plotCellMiner2D(df, xCol = "x", yCol = "y", xPlotLabel = xCol,
  yPlotLabel = yCol, title = NULL, showColorTissues = TRUE,
  showLegend = TRUE, showTrendLine = TRUE, showTitle = TRUE)
}
\arguments{
\item{df}{a data.frame with at least two columns}

\item{xCol}{the name of the column in df with the "x" data}

\item{yCol}{the name of the column in df with the "y" data}

\item{xPlotLabel}{the x plot label}

\item{yPlotLabel}{the y plot label}

\item{title}{the plot title, if null the correlation will appear (DEFAULT: null)}

\item{showColorTissues}{boolean, whether to show tissue colors}

\item{showLegend}{boolean, whether to show the legend}

\item{showTrendLine}{boolean, whether to show the trendline}

\item{showTitle}{NNNNN}
}
\value{
a ggplot object
}
\description{
Make a simple 2d plot using two variables with ggplot2
}
\examples{
\dontrun{
# Load data
nci60DrugActZ <- exprs(getAct(rcellminerData::drugData))
nci60GeneExpZ <- getAllFeatureData(rcellminerData::molData)[["exp"]]
# Load colors
colorTab <- loadNciColorSet(returnDf=TRUE)
tissueColorTab <- unique(colorTab[, c("tissues", "colors")])
# Merge data
df <- as.data.frame(t(rbind(nci60DrugActZ["94600",], nci60GeneExpZ["SLFN11",])))
colnames(df) <- c("y", "x")
df <- cbind(df, colorTab)
# Plot data
plotCellMiner2D(df, xCol="x", yCol="y", xPlotLabel="SLFN11", yPlotLabel="94600")
plotCellMiner2D(df, xCol="x", yCol="y", showTrendLine = FALSE, showTitle = FALSE)
plotCellMiner2D(df, xCol="x", yCol="y", showTrendLine = FALSE, showLegend = FALSE)
}

}
\author{
Augustin Luna <augustin AT mail.nih.gov>
}
\concept{
rcellminer
}

