% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStructures.R
\name{plotStructures}
\alias{plotStructures}
\title{Plot Structures}
\usage{
plotStructures(ids, smiles, titleCex = 1, structSize = 300,
  structAnnotPos = 50, mainLabel = "", rows = 1, cols = length(ids))
}
\arguments{
\item{ids}{a vector of strings of IDs used as structure titles}

\item{smiles}{a vector of strings where the strings are SMILES structures}

\item{titleCex}{a number, the scaling factor for the title (default: 1)}

\item{structSize}{a number, the size of the structure image (default: 200)}

\item{structAnnotPos}{a number, how far above the structure to display the title (default: 50)}

\item{mainLabel}{a string, the main plot label}

\item{rows}{number of rows in figure (default: 1)}

\item{cols}{number of columns in figure (default: input structures number)}
}
\value{
the function does not return anything
}
\description{
Plot Structures
}
\details{
The parameter ids is used as a title, this function does not search for IDs, but works
  based off the smiles given. This is a wrapper around rcdkplot for plotting multiple structures.
}
\examples{
drugAnnot <- as(featureData(getAct(rcellminerData::drugData)), "data.frame")
plotStructures("94600", drugAnnot["94600","SMILES"])
plotStructures(c("609699", "94600"), drugAnnot[c("609699", "94600"),"SMILES"],
  mainLabel=c("609699", "94600"))
}
\author{
Augustin Luna <augustin AT mail.nih.gov>
}
\concept{
rcellminer
}

