% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdkplot.R
\name{rcdkplot}
\alias{rcdkplot}
\title{Plot molecules}
\usage{
rcdkplot(molecule, width = 300, height = 300, marg = 0, main = "")
}
\arguments{
\item{molecule}{an RCDK molecule}

\item{width}{an integer width of the molecule}

\item{height}{an integer height of the molecule}

\item{marg}{margin for all side of the plot (default: 0)}

\item{main}{a string the main title of the figure}
}
\value{
None
}
\description{
Plot molecules in R plot window instead of separate Java window
}
\details{
Code taken from: 
http://www.cureffi.org/2013/09/23/a-quick-intro-to-chemical-informatics-in-r/
}
\examples{
tmp <- parse.smiles("C1=CN(C(=O)N=C1N)C2C(C(C(O2)CO)O)(F)F")
rcdkplot(tmp[[1]], width=300, height=300, main="Gemcitabine")

}
\concept{
rcellminer
}

