% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMolDataType.R
\name{removeMolDataType}
\alias{removeMolDataType}
\title{Remove molecular data type prefixes from features.}
\usage{
removeMolDataType(features, prefixLen = 3)
}
\arguments{
\item{features}{A vector of features.}

\item{prefixLen}{The length of the molecular data type prefix.}
}
\value{
A named vector of features without molecular data type prefixes.
}
\description{
Remove molecular data type prefixes from features.
}
\details{
This function is primarily used to remove prefixes from elastic 
  net features.
}
\examples{
removeMolDataType(c("expTP53", "copMDM2", "mutCHEK2", "mutBRAF"))

}
\concept{rcellminer}
