% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowCors.R
\name{rowCors}
\alias{rowCors}
\title{Row-wise correlations}
\usage{
rowCors(X, Y)
}
\arguments{
\item{X}{a matrix}

\item{Y}{a matrix}
}
\value{
a list of two vectors: cor (correlation values) and pval (correlation p-values)
}
\description{
Correlation between ith row of x and ith row of y for all i
}
\examples{
a <- matrix(runif(100), nrow=10, ncol=10)
b <- matrix(runif(100), nrow=10, ncol=10)
c <- rowCors(a, b)
}
\author{
Sudhir Varma, NCI-LMP
}
\concept{
rcellminer
}

