% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyComparePlots.R
\name{runShinyComparePlots}
\alias{runShinyComparePlots}
\title{Run the Compare Plots Shiny App}
\usage{
runShinyComparePlots(launch.browser = TRUE, port = 3838)
}
\arguments{
\item{launch.browser}{launch in browser? (default: TRUE)}

\item{port}{port number to use}
}
\value{
None
}
\description{
Run the Compare Plots Shiny App
}
\examples{
port <-3838
# Uncomment first
#runShinyComparePlots(port=port)

}
\concept{rcellminer}
