% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyCompareStructures.R
\name{runShinyCompareStructures}
\alias{runShinyCompareStructures}
\title{Run the Compare Structures Shiny App}
\usage{
runShinyCompareStructures(launch.browser = TRUE, port = 3838)
}
\arguments{
\item{launch.browser}{launch in browser? (default: TRUE)}

\item{port}{port number to use}
}
\value{
None
}
\description{
Run the Compare Structures Shiny App
}
\examples{
port <- 3838
# Uncomment first
#runShinyCompareStructures(port=port)

}
\concept{
rcellminer
}
