% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyCompoundBrowser.R
\name{runShinyCompoundBrowser}
\alias{runShinyCompoundBrowser}
\title{Run the Compound Browser}
\usage{
runShinyCompoundBrowser(launch.browser = TRUE, port = 3838)
}
\arguments{
\item{launch.browser}{launch in browser? (default: TRUE)}

\item{port}{port number to use}
}
\value{
None
}
\description{
Run the Compound Browser
}
\examples{
port <- 3838
# Uncomment first
#runShinyCompoundBrowser(port=port)

}
\concept{rcellminer}
