% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchForNscs.R
\name{searchForNscs}
\alias{searchForNscs}
\title{Search for NSCs}
\usage{
searchForNscs(pattern)
}
\arguments{
\item{pattern}{a search pattern. This string will be treated as a regular expression
with the case ignored.}
}
\value{
A vector of matching NSCs
}
\description{
Search for NSCs
}
\details{
Use this function with caution. Not all compounds have names and
  compounds can have many synonyms not included in CellMiner.
}
\examples{
searchForNscs("nib$")
}
\concept{
rcellminer
}

