% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectCorrelatedRows2.R
\name{selectCorrelatedRows2}
\alias{selectCorrelatedRows2}
\title{Get other features above some threshold (merges results from selectCorrelatedRows,
invoked for a given Y, and each element in a list of matrices).}
\usage{
selectCorrelatedRows2(Y, XList, corThreshold = 0.1, useAbsCor = TRUE)
}
\arguments{
\item{Y}{a vector or matrix; rows from each matrix element of X will be correlated with
Y if Y is a vector or with rows of Y, if Y is a matrix.}

\item{XList}{a list of matrices whose rows will be correlated with Y (vector) or rows of Y (matrix)}

\item{corThreshold}{the minimum correlation threshold for the row to be returned}

\item{useAbsCor}{a logical value indicating whether absolute correlations should be used
(default=TRUE).}
}
\value{
a matrix formed from rows of matrices in XList that are correlated with Y
(if Y is a vector) or correlated with at least one row of Y if Y is a matrix or data frame.
}
\description{
Get other features above some threshold (merges results from selectCorrelatedRows,
invoked for a given Y, and each element in a list of matrices).
}
\examples{
vec <- runif(10)
names(vec) <- 1:10
matList <- list(X1 = matrix(runif(100), 10, 10), X2 = matrix(runif(100), 10, 10))
rownames(matList$X1) <- paste0("X1_row_", 1:10)
colnames(matList$X1) <- paste0("X1_col_", 1:10)
rownames(matList$X2) <- paste0("X2_row_", 1:10)
colnames(matList$X2) <- paste0("X2_col_", 1:10)
selectCorrelatedRows2(vec, matList)
}
\concept{
rcellminer
}

