% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{[[,MolData-method}
\alias{[[,MolData-method}
\title{Returns an indexed eSet object from a MolData object eSet list.}
\usage{
\S4method{[[}{MolData}(x, i)
}
\arguments{
\item{x}{A MolData object.}

\item{i}{Index or named item in MolData object eSet list.}
}
\value{
An indexed eSet object from a MolData object eSet list.
}
\description{
Returns an indexed eSet object from a MolData object eSet list.
}
\concept{
rcellminerData
}
