% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MolData-methods.R
\docType{methods}
\name{[[<-,MolData-method}
\alias{[[<-,MolData-method}
\title{Assigns an eSet object to a specified position in a MolData object eSet list.}
\usage{
\S4method{[[}{MolData}(x, i) <- value
}
\arguments{
\item{x}{A MolData object.}

\item{i}{Index or named item in MolData object eSet list.}

\item{value}{An eSet object to be assigned.}
}
\value{
An eSet object to a specified position in a MolData object eSet list.
}
\description{
Assigns an eSet object to a specified position in a MolData object eSet list.
}
\concept{
rcellminerData
}

