% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{cmVersion}
\alias{cmVersion}
\title{CellMiner Version}
\description{
CellMiner Version
}
\details{
The version of CellMiner used
}
\author{
Vinodh Rajapakse \email{vinodh.rajapakse AT nih.gov}
}
\references{
\url{http://discover.nci.nih.gov/cellminer}
}
\concept{
rcellminer
}
\keyword{data}

