% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFingerprintList.R
\docType{data}
\name{fingerprintList}
\alias{fingerprintList}
\title{A list of pre-computed fingerprints using getFingeprintList()}
\description{
A list of pre-computed fingerprints using getFingeprintList()
}
\concept{
rcellminer
}
\keyword{data}

