% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStructuresFromNscs.R
\name{plotStructuresFromNscs}
\alias{plotStructuresFromNscs}
\title{Plot the structures for NSCs}
\usage{
plotStructuresFromNscs(nscs)
}
\arguments{
\item{nscs}{a vector of strings of NSCs used as structure titles}
}
\value{
the function does not return anything
}
\description{
Plot the structures for NSCs
}
\details{
This is a wrapper for the plotStructures() function that takes only NSCs
}
\examples{
plotStructuresFromNscs("94600")
plotStructuresFromNscs(c("609699", "94600"))

}
\author{
Augustin Luna <augustin AT mail.nih.gov>
}
\concept{
rcellminer
}
